[dgarciabravo] Resolucion incidencia SIGARNV-1903: Da error al subir los documentos en envios masivos de tipo correo electronico

+++ siga-web/src/main/java/org/itcgae/siga/com/services/impl/EnviosMasivosServiceImpl.java
@@ -2222,7 +2222,7 @@ public class EnviosMasivosServiceImpl implements IEnviosMasivosService {
 	}
 
 	@Override
-	public Resource recuperaPdfBuroSMS(Short idInstitucion, Long idEnvio, Integer idDocumento) {
+	public Resource recuperaPdfBuroSMS(Short idInstitucion, Long idEnvio, Short idDocumento) {
 
 		LOGGER.debug("Comprobando si es un envío burosms para recuperar el pdf para iddocumento " + idDocumento);
 		Resource inputStreamResource = null;
